+++ dev_tools/bash_scripts_test.py
@@ -16,7 +16,7 @@ import os
 from typing import TYPE_CHECKING, Iterable
 
 from dev_tools import shell_tools
-from dev_tools.utils import only_on_posix
+from dev_tools.test_utils import only_on_posix
 
 if TYPE_CHECKING:
     import _pytest.tmpdir
