+++ dev_tools/shell_tools_test.py
@@ -17,7 +17,7 @@ import subprocess
 import pytest
 
 from dev_tools import shell_tools
-from dev_tools.utils import only_on_posix
+from dev_tools.test_utils import only_on_posix
 
 
 def run_cmd(*args, **kwargs):
